+++ templates/header.php
@@ -22,3 +22,4 @@ if ( ! defined( 'ABSPATH' ) ) {
 </head>
 <body <?php body_class(); ?>>
 
+a