+++ CHANGELOG.md
@@ -1,7 +1,35 @@
 # Netivo CLI Changelog
 
+## Version 1.2.8
+- Fixed server side rendered blocks, render_callback error
+
+## Version 1.2.7
+- Fixed server side rendered blocks
+
+## Version 1.2.6
+- Changed edit and save block file generation.
+
+## Version 1.2.5
+- Fixed blok generation no block name in block.json and wrong style directory
+- Added default options to block.json: icon, keywords
+
+## Version 1.2.4
+- Fixed pre-commit
+
+## Version 1.2.3
+- Added phpcs when creating a new project
+- Added git initialization with a pre-commit hook during project creation
+- Changed log to apply colors to error and waring logs.
+
+## Version 1.2.2
+- Generate theme.json during project creation
+
+## Version 1.2.1
+- Composer modules version change
+
 ## Version 1.2
 - Added action for creating gutenberg blocks
+- Changed project generation to use new core modules
 
 ## Version 1.1
 - Added action for creating metaboxes